[actions] Add mongoDB to build&test checks  (#40)
* [actions] Add mongodb step to build&test

* [tests] Unskip recipesController tests

* [tests] Fix dropping collection when doesnt exist

* Reduce number of node versions in built-test workflow

+++ server/test/utils.js
@@ -3,7 +3,12 @@
 const mongoose = require('mongoose');
 
 const initEmptyDatabase = async () => {
-    await mongoose.connection.db.dropCollection('recipes');
+    const { db } = mongoose.connection;
+
+    const currentCollections = await db.listCollections({ name: 'recipes' });
+    if (currentCollections.readableLength > 0) {
+        await mongoose.connection.db.dropCollection('recipes');
+    }
     await mongoose.connection.db.createCollection('recipes');
 };
